+++ wagtail_content_import/pickers/local/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailContentImportLocalPickerAppConfig(AppConfig):
+    name = 'wagtail_content_import.pickers.local'
+    label = 'wagtail_content_import_local_picker'
+    verbose_name = "Wagtail Content Import - Local File Picker"